mlops-project/src/requirements.txt
numpy
scikit-learn
azureml-inference-server-http
pandas
azureml-sdk
sklearn